+++ .env
@@ -1,0 +1,1 @@
+VITE_BACKEND_URL = http://127.0.0.1:8000/api/v1/category_grafic/
