+++ package.json
@@ -14,12 +14,10 @@
     "@mui/material": "^5.11.15",
     "@react-spring/web": "^9.7.1",
     "@types/victory": "^35.0.0",
-    "axios": "^1.3.4",
     "chart.js": "^4.2.1",
     "react": "^18.2.0",
     "react-chartjs-2": "^5.2.0",
-    "react-dom": "^18.2.0",
-    "victory": "^36.6.8"
+    "react-dom": "^18.2.0"
   },
   "devDependencies": {
     "@types/react": "^18.0.28",
