+++ src/components/App/AddNewCategory.tsx
@@ -1,0 +1,2 @@
+import Modal from '@mui/material/Modal';
+export default Modal;
