+++ src/components/App/RandomGrafic.tsx
@@ -1,7 +1,4 @@
 import React, {useLayoutEffect, useEffect, useState} from 'react'
-import axios from 'axios'
-
-
 import {Chart as ChartJS, BarElement, CategoryScale, LinearScale, Tooltip, Legend} from 'chart.js'
 import {Bar} from 'react-chartjs-2'
 
@@ -23,18 +20,18 @@ const RandomGrafic = () => {
     
         }]
     })
+
     useEffect(() => {
         const fetchData = async() =>{
-            const dataset = [];
-            await fetch("http://127.0.0.1:8000/api/v1/category_grafic/")
+            const dataset: any[] = [];
+            await fetch(import.meta.env.VITE_BACKEND_URL)
             .then(data => {
                 const res = data.json()
                 return res
                 
             }).then((res) =>{
-                console.log(res.category_grafic_1)
-                for(const val of Object.keys(res.category_grafic_1)){
-                    dataset.push(res.category_grafic_1[val].price)
+                for(const val of Object.keys(res.CategoryGrafic1)){
+                    dataset.push(res.CategoryGrafic1[val].price)
                 }
                 setData({labels: ['27.01', '26.01', '25.01'],
                 datasets:[{
@@ -44,7 +41,6 @@ const RandomGrafic = () => {
                     borderWidth:1,
             
                 }]})
-                //console.log(dataset)
             })
             .catch(error => {
                 console.error(error);
@@ -55,30 +51,6 @@ const RandomGrafic = () => {
     },[])
 
     
-
-    // async function getData() {
-    //     const response = await fetch("http://127.0.0.1:8000/api/v1/category_grafic/");
-    //     const fetch_data = await response.json();
-    //     console.log(fetch_data);
-    //     return fetch_data;
-    //     }
-    //     getData().then(fetch_data =>{
-    //         const price = fetch_data.category_grafic_1.map(function(index){
-    //             return index.price
-    //         })
-    //         console.log(price)
-    //     })
-
-    // const data = {
-    //     labels: ['27.01', '26.01', '25.01'],
-    //     datasets:[{
-    //         data: g_data,
-    //         backgroundColor: 'black',
-    //         borderColor: 'black',
-    //         borderWidth:1,
-
-    //     }]
-    // }
     const options ={
         plugins: {
             legend: false
@@ -87,9 +59,6 @@ const RandomGrafic = () => {
         maintainAspectRatio: false,
         //barThickness:20
     }
-
-
-
     return(
         <div className="random_grafic">
              <div className="random_grafic_cont">
