+++ src/components/App/RecentBoughts.tsx
