--- src/components/App/add_new_category.tsx
@@ -1,13 +1,0 @@
-import React from 'react';
-
-
-
-const Modal = ({active, setActive}) =>{
-    return (
-        <div className={active ? "active" : "non_active"}  onClick={() => setActive(false)}>
-            <p>Modal</p>
-        </div>
-    )
-}
-
-export default Modal
