+++ src/components/App/index.tsx
@@ -1,11 +1,11 @@
 import React, {Component, Fragment, useState} from 'react'
-import MainPage from './mainpage'
-import Recent from './recent_boughts'
-import RandomGrafic from './random_grafic'
-import Sum from './sum'
+
+import Recent from './RecentBoughts'
+import RandomGrafic from './RandomGrafic'
+import Sum from './Sum'
 import imgSrc from '../../assets/logo.png';
-import Modal from './add_new_category'
-import Chart from './d'
+import Modal from './AddNewCategory'
+import Chart from './Chart'
 
 
 const App = () => {
@@ -20,45 +20,21 @@ const App = () => {
         </div>
         
         <div className="grafics">
-
-          <div className="random_grafics">
-            <Fragment>
-              <RandomGrafic />
-            </Fragment>
-
-            <Fragment>
-              <RandomGrafic />
-            </Fragment>
-
-            <Fragment>
-              <RandomGrafic />
-            </Fragment>
-          </div>
-
           <div className="pieGrafic">
-            <Fragment >
-              <Chart />
-            </Fragment>
+            <Chart />
           </div>
-
+          <div className="random_grafics">
+              <RandomGrafic />                
+              <RandomGrafic />
+          </div>
         </div>
-
       </div>
 
       <div className="right">
       <img className='logo' src={imgSrc} alt="" />
-          <Fragment>
-            <Sum />
-          </Fragment>
-          <Fragment>
-            <Modal active={modalActive} setActive={setModalActive}/>
-          </Fragment> 
+          <Sum />
           <button className='add_new_entry_btn' onClick={() => setModalActive(true)}>Add New Entry</button>
-          <Fragment>
-            <Recent />
-          </Fragment>
-
-
+          <Recent />
       </div>
     </div>
     )
