+++ src/components/App/sum.tsx
@@ -1,4 +1,4 @@
-import React, {useLayoutEffect} from 'react'
+import React from 'react'
 
 const Sum = () => {
     return (
